Move performance tests to own folder
