+++ tests/integration/test_integration_space_time.py
@@ -247,49 +247,3 @@ def test_integration_space_time_named_node_set(
         get_corresponding_reference_file_path(extension="json"),
         mesh_data_arrays,
     )
-
-
-@pytest.mark.performance
-def test_integration_space_time_performance_create_mesh_in_space(
-    evaluate_execution_time, cache_data
-):
-    """Test the performance of the mesh creation in space."""
-
-    mesh = Mesh()
-    beam_type = generate_beam_class(3)
-
-    evaluate_execution_time(
-        "BeamMe: Space-Time: Create mesh in space",
-        create_beam_mesh_line,
-        kwargs={
-            "mesh": mesh,
-            "beam_class": beam_type,
-            "material": MaterialBeamBase(),
-            "start_point": [0, 0, 0],
-            "end_point": [1, 0, 0],
-            "n_el": 100,
-        },
-        expected_time=0.01,
-    )
-
-    # store mesh in cache for upcoming test
-    cache_data.mesh = mesh
-
-
-@pytest.mark.performance
-def test_integration_space_time_performance_create_mesh_in_time(
-    evaluate_execution_time, cache_data
-):
-    """Test the performance of the mesh creation in time."""
-
-    evaluate_execution_time(
-        "BeamMe: Space-Time: Create mesh in time",
-        beam_to_space_time,
-        kwargs={
-            "mesh_space_or_generator": cache_data.mesh,
-            "time_duration": 6.9,
-            "number_of_elements_in_time": 1000,
-            "time_start": 1.69,
-        },
-        expected_time=3.0,
-    )
