+++ tests/performance/test_performance_beamme.py
@@ -175,7 +175,7 @@ def create_beam_mesh(n_x, n_y, n_z, n_el):
 
 
 @pytest.mark.performance
-def test_integration_performance_cubitpy_create_solid(
+def test_performance_beamme_cubitpy_create_solid(
     evaluate_execution_time, shared_tmp_path
 ):
     """Test the performance of creating a solid block using CubitPy.
@@ -201,7 +201,7 @@ def test_integration_performance_cubitpy_create_solid(
 
 
 @pytest.mark.performance
-def test_integration_performance_double_solid_block(
+def test_performance_beamme_double_solid_block(
     evaluate_execution_time, shared_tmp_path
 ):
     """The version of Cubit we use in testing only allows for 50,000 elements,
@@ -236,7 +236,7 @@ def test_integration_performance_double_solid_block(
     ],
 )
 @pytest.mark.performance
-def test_integration_performance_load_solid(
+def test_performance_beamme_load_solid(
     evaluate_execution_time, shared_tmp_path, log_name, full_import, expected_time
 ):
     """Test the performance of loading a solid mesh."""
@@ -253,7 +253,7 @@ def test_integration_performance_load_solid(
 
 
 @pytest.mark.performance
-def test_integration_performance_create_beams(evaluate_execution_time, cache_data):
+def test_performance_beamme_create_beams(evaluate_execution_time, cache_data):
     """Test the performance of creating a large beam mesh."""
 
     # store mesh in cache for upcoming tests
@@ -271,7 +271,7 @@ def test_integration_performance_create_beams(evaluate_execution_time, cache_dat
 
 
 @pytest.mark.performance
-def test_integration_performance_rotate(evaluate_execution_time, cache_data):
+def test_performance_beamme_rotate(evaluate_execution_time, cache_data):
     """Test the performance of rotating a large beam mesh."""
 
     evaluate_execution_time(
@@ -283,7 +283,7 @@ def test_integration_performance_rotate(evaluate_execution_time, cache_data):
 
 
 @pytest.mark.performance
-def test_integration_performance_translate(evaluate_execution_time, cache_data):
+def test_performance_beamme_translate(evaluate_execution_time, cache_data):
     """Test the performance of translating a large beam mesh."""
 
     evaluate_execution_time(
@@ -295,7 +295,7 @@ def test_integration_performance_translate(evaluate_execution_time, cache_data):
 
 
 @pytest.mark.performance
-def test_integration_performance_reflect(evaluate_execution_time, cache_data):
+def test_performance_beamme_reflect(evaluate_execution_time, cache_data):
     """Test the performance of reflecting a large beam mesh."""
 
     evaluate_execution_time(
@@ -307,7 +307,7 @@ def test_integration_performance_reflect(evaluate_execution_time, cache_data):
 
 
 @pytest.mark.performance
-def test_integration_performance_mespy_wrap_around_cylinder(
+def test_performance_beamme_mespy_wrap_around_cylinder(
     evaluate_execution_time, cache_data
 ):
     """Test the performance of wrapping a large beam mesh around a cylinder."""
@@ -321,7 +321,7 @@ def test_integration_performance_mespy_wrap_around_cylinder(
 
 
 @pytest.mark.performance
-def test_integration_performance_wrap_around_cylinder_without_check(
+def test_performance_beamme_wrap_around_cylinder_without_check(
     evaluate_execution_time, cache_data
 ):
     """Test the performance of wrapping a large beam mesh around a cylinder
@@ -336,7 +336,7 @@ def test_integration_performance_wrap_around_cylinder_without_check(
 
 
 @pytest.mark.performance
-def test_integration_performance_find_close_nodes(evaluate_execution_time, cache_data):
+def test_performance_beamme_find_close_nodes(evaluate_execution_time, cache_data):
     """Test the performance of finding close nodes in a large beam mesh."""
 
     evaluate_execution_time(
@@ -348,9 +348,7 @@ def test_integration_performance_find_close_nodes(evaluate_execution_time, cache
 
 
 @pytest.mark.performance
-def test_integration_performance_add_mesh_to_input_file(
-    evaluate_execution_time, cache_data
-):
+def test_performance_beamme_add_mesh_to_input_file(evaluate_execution_time, cache_data):
     """Test the performance of adding a mesh to an input file."""
 
     input_file = InputFile()
@@ -366,7 +364,7 @@ def test_integration_performance_add_mesh_to_input_file(
 
 
 @pytest.mark.performance
-def test_integration_performance_dump_input_file(
+def test_performance_beamme_dump_input_file(
     evaluate_execution_time,
     cache_data,
     tmp_path,
@@ -385,9 +383,7 @@ def test_integration_performance_dump_input_file(
 
 
 @pytest.mark.performance
-def test_integration_performance_write_vtk(
-    evaluate_execution_time, tmp_path, cache_data
-):
+def test_performance_beamme_write_vtk(evaluate_execution_time, tmp_path, cache_data):
     """Test the performance of writing a beam mesh to VTK format."""
 
     # use a smaller mesh for testing vtk output performance
@@ -406,7 +402,7 @@ def test_integration_performance_write_vtk(
 
 
 @pytest.mark.performance
-def test_integration_performance_write_vtk_smooth(
+def test_performance_beamme_write_vtk_smooth(
     evaluate_execution_time, tmp_path, cache_data
 ):
     """Test the performance of writing a beam mesh to VTK format with more
